[PI-15] Fix month core and change file location
