+++ AppPackage/Sources/CalendarFeature/Month/MonthView.swift
@@ -45,38 +45,31 @@ struct MonthView: View {
             ),
             spacing: .zero
         ) {
-            ForEach(viewStore.days) { day in
-                Text(day.date.dayString)
-                    .bold(day.isToday)
-                    .foregroundColor(.dayColor(date: day.date, isFaded: day.isFaded))
-                    .frame(height: layoutConstraint.rowHeight)
-                    .frame(maxWidth: .infinity)
-                    .background {
-                        ZStack {
-                            switch day.selectionType {
-                            case .clear:
-                                Color.clear
-                            case .painted:
-                                PDS.COLOR.purple9.scale
-                                    .cornerRadius(11)
-                                    .frame(width: 32, height: 30)
-                            }
-                        }
-                    }
-                    .onTapGesture { viewStore.send(.rowTapped(day.date)) }
-                    .disabled(type == .appointment)
-            }
+            ForEachStore(
+                store
+                    .scope(
+                        state: \.monthState.dayStateList,
+                        action: { MonthCore.Action.delegate(action: .day(id: $0, action: $1)) }
+                    ),
+                content: {
+                    DayView(
+                        calendarType: type,
+                        layoutConstraint: .init(rowHeight: layoutConstraint.rowHeight),
+                        store: $0
+                    )
+                }
+            )
         }
         .id(viewStore.id)
         .gesture(
             DragGesture(
                 minimumDistance: .zero,
-                coordinateSpace: type == .appointment
+                coordinateSpace: type == .promise
                     ? .local
                     : .global
             )
             .onChanged {
-                guard type == .appointment else { return }
+                guard type == .promise else { return }
                 let startIndex = transformToIndex(
                     point: $0.startLocation,
                     viewWidth: geometryWidth
@@ -88,7 +81,7 @@ struct MonthView: View {
                 viewStore.send(.drag(startIndex: startIndex, endIndex: endIndex))
             }
             .onEnded {
-                guard type == .appointment else { return }
+                guard type == .promise else { return }
                 let startIndex = transformToIndex(
                     point: $0.startLocation,
                     viewWidth: geometryWidth
@@ -111,14 +104,14 @@ struct MonthView: View {
 private extension MonthView {
     struct ViewState: Equatable {
         let id: Date
-        let days: [Day]
+        let days: IdentifiedArrayOf<DayCore.State>
     }
 
     enum ViewAction: Equatable {
         var reducerAction: MonthCore.Action {
             switch self {
-            case let .rowTapped(date):
-                return .delegate(action: .rowTapped(date))
+            case let .day(id: id, action: dayAction):
+                return .delegate(action: .day(id: id, action: dayAction))
 
             case let .drag(startIndex: startIndex, endIndex: endIndex):
                 return .delegate(action: .drag(startIndex: startIndex, endIndex: endIndex))
@@ -128,7 +121,7 @@ private extension MonthView {
             }
         }
 
-        case rowTapped(Date)
+        case day(id: DayCore.State.ID, action: DayCore.Action)
         case drag(startIndex: Int, endIndex: Int)
         case dragEnded(startIndex: Int)
     }
@@ -136,24 +129,6 @@ private extension MonthView {
 
 private extension MonthCore.State {
     var viewState: MonthView.ViewState {
-        .init(id: id, days: monthState.days)
-    }
-}
-
-private extension Date {
-    var dayString: String {
-        formatted(
-            .dateTime
-                .day()
-        )
-    }
-}
-
-private extension Color {
-    static func dayColor(date: Date, isFaded: Bool) -> Self {
-        guard !isFaded else { return PDS.COLOR.gray3.scale }
-        return calendar.component(.weekday, from: date) == 1
-            ? PDS.COLOR.scarlet1.scale
-            : PDS.COLOR.cGray2.scale
+        .init(id: id, days: monthState.dayStateList)
     }
 }
